C# Example structure updated with API 3.2.1
